+++ Interfaces/IHealth.cs
@@ -2,12 +2,12 @@
 {
     public interface IHealth
     {
-        public int thorax_health { get; set; }
-        public int stomach_health { get; set; }
-        public int head_health { get; set; }
-        public int left_arm_health { get; set; }
-        public int right_arm_health { get; set; }
-        public int left_leg_health { get; set; }
-        public int right_leg_health { get; set; }
+        public float thorax_health { get; set; }
+        public float stomach_health { get; set; }
+        public float head_health { get; set; }
+        public float left_arm_health { get; set; }
+        public float right_arm_health { get; set; }
+        public float left_leg_health { get; set; }
+        public float right_leg_health { get; set; }
     }
 }
